Add from-zero-to-rapids.tex and from-zero-to-rapids.pdf
XSEDE Webinar given on Thursday, January 23rd, 2020. From Zero to
RAPIDS: Accelerating Data Science and Machine Learning Workflows on
NVIDIA GPUs. RAPIDS is NVIDIA's new Python-based framework for
accelerating end-to-end data science and machine learning pipelines
on their GPUs. In this webinar, we'll provide an overview of this new
framework and how you can incorporate it in your own research. In
particular, if you use pandas and/or scikit-learn extensively, you'll
definitely want to attend this webinar.